# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AviFit APP  API",
    "description": "This is the documentation for my API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/categories": {
      "get": {
        "tags": ["Category"],
        "summary": "Get all categories",
        "description": "Retrieve a list of all categories.",
        "responses": {
          "200": {
            "description": "A list of categories.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Category"],
        "summary": "Create a new category",
        "description": "Create a new category.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "tags": ["Category"],
        "summary": "Get a category by ID",
        "description": "Retrieve a specific category by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the category to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Category details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "put": {
        "tags": ["Category"],
        "summary": "Update a category by ID",
        "description": "Update an existing category.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the category to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "delete": {
        "tags": ["Category"],
        "summary": "Delete a category by ID",
        "description": "Delete a specific category by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the category to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted successfully"
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
    "/api/subcategories": {
      "get": {
        "tags": ["Subcategory"],
        "summary": "Get all subcategories",
        "description": "Retrieve a list of all subcategories.",
        "responses": {
          "200": {
            "description": "A list of subcategories.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subcategory"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Subcategory"],
        "summary": "Create a new subcategory",
        "description": "Create a new subcategory.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubcategory"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subcategory created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subcategory"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/tutorials": {
      "get": {
        "tags": ["Tutorial"],
        "summary": "Get all tutorials",
        "description": "Retrieve a list of all tutorials.",
        "responses": {
          "200": {
            "description": "A list of tutorials.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tutorial"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/videos": {
      "get": {
        "tags": ["Video"],
        "summary": "Get all videos",
        "description": "Retrieve a list of all videos.",
        "responses": {
          "200": {
            "description": "A list of videos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        }
      },
      "CreateCategory": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fitness"
          },
          "color": {
            "type": "string",
            "example": "#ff0000"
          }
        }
      },
      "Subcategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer"
          }
        }
      },
      "CreateSubcategory": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Cardio"
          },
          "color": {
            "type": "string",
            "example": "#00ff00"
          },
          "categoryId": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "Tutorial": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "tutorial_code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "video_url": {
            "type": "string"
          },
          "pdf_url": {
            "type": "string"
          }
        }
      },
      "Video": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "video_code": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "video_url": {
            "type": "string"
          },
          "pdf_url": {
            "type": "string"
          }
        }
      }
    }
  }
}